Extract editorconfig related tests to separate file.
