+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/KtlintPluginTest.kt
@@ -23,19 +23,7 @@ abstract class BaseKtlintPluginTest : AbstractPluginTest() {
 
     @BeforeEach
     fun setupBuild() {
-        projectRoot.apply {
-            buildFile().writeText("""
-                ${pluginsBlockWithMainPluginAndKotlinJvm()}
-
-                repositories {
-                    gradlePluginPortal()
-                }
-
-                import org.jlleitschuh.gradle.ktlint.reporter.ReporterType
-
-                ktlint.reporters = [ReporterType.CHECKSTYLE, ReporterType.PLAIN]
-            """.trimIndent())
-        }
+        projectRoot.defaultProjectSetup()
     }
 
     @Test
@@ -130,85 +118,6 @@ abstract class BaseKtlintPluginTest : AbstractPluginTest() {
             // TODO: Stale reports are not cleaned up
             assertReportCreated(ReporterType.PLAIN)
         }
-    }
-
-    @Test
-    fun `Check task should be up_to_date if editorconfig content not changed`() {
-        projectRoot.withCleanSources()
-        projectRoot.createEditorconfigFile()
-
-        build("ktlintCheck").apply {
-            assertThat(task(":ktlintMainSourceSetCheck")!!.outcome).isEqualTo(TaskOutcome.SUCCESS)
-        }
-        build("ktlintCheck").apply {
-            assertThat(task(":ktlintMainSourceSetCheck")!!.outcome).isEqualTo(TaskOutcome.UP_TO_DATE)
-        }
-    }
-
-    @Test
-    fun `Check task should rerun if editorconfig content changed`() {
-        projectRoot.withCleanSources()
-        projectRoot.createEditorconfigFile()
-
-        build("ktlintCheck").apply {
-            assertThat(task(":ktlintMainSourceSetCheck")!!.outcome).isEqualTo(TaskOutcome.SUCCESS)
-        }
-
-        projectRoot.modifyEditorconfigFile(100)
-        build("ktlintCheck").apply {
-            assertThat(task(":ktlintMainSourceSetCheck")!!.outcome).isEqualTo(TaskOutcome.SUCCESS)
-        }
-    }
-
-    @Test
-    internal fun `Check task should rerun if root editorconfig content changed`() {
-        val projectWithModulesLocation = temporaryFolder.resolve("modularized")
-        projectWithModulesLocation.mkdirs()
-        val moduleLocation = projectWithModulesLocation.resolve("test/module1")
-        moduleLocation.mkdirs()
-
-        projectWithModulesLocation.settingsFile().writeText("""
-            include ":test:module1"
-        """.trimIndent())
-        projectWithModulesLocation.buildFile().writeText("""
-            ${pluginsBlockWithMainPluginAndKotlinJvm()}
-
-            repositories {
-                gradlePluginPortal()
-            }
-
-            allprojects {
-                repositories {
-                    jcenter()
-                }
-            }
-        """.trimIndent())
-        projectWithModulesLocation.createEditorconfigFile()
-        moduleLocation.buildFile().writeText("""
-            apply plugin: "kotlin"
-            apply plugin: "org.jlleitschuh.gradle.ktlint"
-        """.trimIndent())
-        moduleLocation.withCleanSources()
-
-        val gradleRunner = GradleRunner.create()
-            .withProjectDir(projectWithModulesLocation)
-            .withPluginClasspath()
-
-        gradleRunner
-            .withArguments(":test:module1:ktlintCheck")
-            .forwardOutput()
-            .build().apply {
-                assertThat(task(":test:module1:ktlintMainSourceSetCheck")!!.outcome).isEqualTo(TaskOutcome.SUCCESS)
-            }
-
-        projectWithModulesLocation.modifyEditorconfigFile(100)
-
-        gradleRunner
-            .withArguments(":test:module1:ktlintCheck")
-            .forwardOutput()
-            .build().apply {
-                assertThat(task(":test:module1:ktlintMainSourceSetCheck")!!.outcome).isEqualTo(TaskOutcome.SUCCESS)
-            }
     }
 
     @Test
